修改部门相关问题

+++ src/main/java/com/deepsoft/haolifa/service/impl/DepartmentServiceImpl.java
@@ -10,6 +10,7 @@ import com.deepsoft.haolifa.service.DepartmentService;
 import com.deepsoft.haolifa.util.TreeUtils;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -26,6 +27,15 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Override
     public int insertDepartment(DepartmentDTO departmentDTO) {
+        String deptNo = departmentDTO.getDeptNo();
+        if (StringUtils.isNotBlank(deptNo)) {
+            SysDepartmentExample example = new SysDepartmentExample();
+            example.or().andDeptNoEqualTo(deptNo);
+            List<SysDepartment> sysDepartments = departmentMapper.selectByExample(example);
+            if (sysDepartments.size() > 0) {
+                throw new BaseException(CommonEnum.ResponseEnum.FAIL.code, "已经存在相同的deptNo,请换个");
+            }
+        }
         SysDepartment sysDepartment = new SysDepartment();
         BeanUtils.copyProperties(departmentDTO, sysDepartment);
         return departmentMapper.insertSelective(sysDepartment);
@@ -43,6 +53,18 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Override
     public int updateDepartment(DepartmentDTO departmentDTO) {
+        SysDepartmentExample example = new SysDepartmentExample();
+        String deptNo = departmentDTO.getDeptNo();
+        if (StringUtils.isNotBlank(deptNo)) {
+            example.or().andDeptNoEqualTo(deptNo);
+            List<SysDepartment> sysDepartments = departmentMapper.selectByExample(example);
+            if (sysDepartments.size() > 0) {
+                SysDepartment sysDepartment = sysDepartments.get(0);
+                if (sysDepartment.getId() != departmentDTO.getId()) {
+                    throw new BaseException(CommonEnum.ResponseEnum.FAIL.code, "已经存在相同的deptNo,请换个");
+                }
+            }
+        }
         SysDepartment sysDepartment = new SysDepartment();
         BeanUtils.copyProperties(departmentDTO, sysDepartment);
         return departmentMapper.updateByPrimaryKeySelective(sysDepartment);
@@ -52,6 +74,12 @@ public class DepartmentServiceImpl implements DepartmentService {
     public int deleteDepartment(Integer id) {
         if (null == id) {
             throw new BaseException(CommonEnum.ResponseEnum.PARAM_ERROR.code, "id不能为空");
+        }
+        SysDepartmentExample example = new SysDepartmentExample();
+        example.or().andPidEqualTo(id);
+        List<SysDepartment> sysDepartments = departmentMapper.selectByExample(example);
+        if (sysDepartments.size() > 0) {
+            throw new BaseException(CommonEnum.ResponseEnum.FAIL.code, "该部门还有子部门,不能删除");
         }
         return departmentMapper.deleteByPrimaryKey(id);
     }
@@ -81,8 +109,9 @@ public class DepartmentServiceImpl implements DepartmentService {
             departmentTree.setId(String.valueOf(e.getId()));
             departmentTree.setName(e.getDeptName());
             departmentTree.setParentId(String.valueOf(e.getPid()));
+            departmentTree.setNo(e.getDeptNo());
+            departmentTree.setDescription(e.getDescription());
             departmentTrees.add(departmentTree);
-
         });
         return TreeUtils.getTreeList("0", departmentTrees);
     }
